SSODefinition form & fix path
